// src/types/ClassicType.ts
export type ClassicCertData = {
  owner: string;
  boatName: string;
  className?: string;
  division: string;
  subDivision?: string;
  sailNumber?: string;
  nonSpinnaker?: number;
  nonSpinnakerTA?: number;
  spinnaker?: number;
  spinnakerTA?: number;
  type: string;
  recorded?: string;
  certId: string;
};